+++ include/dom/events/ui_event.h
@@ -22,24 +22,24 @@ dom_exception _dom_ui_event_get_view(dom_ui_event *evt,
 		(dom_ui_event *) (e), (struct dom_abstract_view **) (v))
 
 dom_exception _dom_ui_event_get_detail(dom_ui_event *evt,
-		long *detail);
+		int32_t *detail);
 #define dom_ui_event_get_detail(e, d) _dom_ui_event_get_detail(\
-		(dom_ui_event *) (e), (long *) (d))
+		(dom_ui_event *) (e), (int32_t *) (d))
 
 dom_exception _dom_ui_event_init(dom_ui_event *evt, dom_string *type, 
 		bool bubble, bool cancelable, struct dom_abstract_view *view,
-		long detail);
+		int32_t detail);
 #define dom_ui_event_init(e, t, b, c, v, d) _dom_ui_event_init( \
 		(dom_ui_event *) (e), (dom_string *) (t), (bool) (b), \
-		(bool) (c), (struct dom_abstract_view *) (v), (long) (d))
+		(bool) (c), (struct dom_abstract_view *) (v), (int32_t) (d))
 
 dom_exception _dom_ui_event_init_ns(dom_ui_event *evt,
 		dom_string *namespace, dom_string *type,
 		bool bubble, bool cancelable, struct dom_abstract_view *view,
-		long detail);
+		int32_t detail);
 #define dom_ui_event_init_ns(e, n, t, b, c, v, d) _dom_ui_event_init_ns( \
 		(dom_ui_event *) (e), (dom_string *) (n), \
 		(dom_string *) (t), (bool) (b), (bool) (c), \
-		(struct dom_abstract_view *) (v), (long) (d))
+		(struct dom_abstract_view *) (v), (int32_t) (d))
 
 #endif
