+++ include/dom/html/html_options_collection.h
@@ -16,11 +16,11 @@ struct dom_node;
 typedef struct dom_html_options_collection dom_html_options_collection;
 
 dom_exception dom_html_options_collection_get_length(
-		dom_html_options_collection *col, unsigned long *len);
+		dom_html_options_collection *col, uint32_t *len);
 dom_exception dom_html_options_collection_set_length(
-		dom_html_options_collection *col, unsigned long len);
+		dom_html_options_collection *col, uint32_t len);
 dom_exception dom_html_options_collection_item(
-		dom_html_options_collection *col, unsigned long index,
+		dom_html_options_collection *col, uint32_t index,
 		struct dom_node **node);
 dom_exception dom_html_options_collection_named_item(
 		dom_html_options_collection *col, dom_string *name,
